fix problem with auto login/refresh
